Move the select_builder command to virtualenv.core
